requirements.txt
##### Dependencies
nulltype
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3,<2.0.0
python-jose

##### DEV Dependencies
cov-core==1.15.0
coverage==5.3
flake8==3.8.4
flake8-mutable==1.2.0
mock==4.0.3
pytest==7.4.2
pytest-xdist==3.3.1
pytest-cov==4.1.0
sphinx==3.3.1
